+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 - Add integration with BigBlueButton video conferencing system !1561 @NerdyProjects
 - Update to Symfony 5 / FOSRestBundle 3 !1573 @NerdyProjects
 - Show an error notification if the user is redirected from a region page to the dashboard !1571 @alex.simm  
+- Show on profile a warning if the private mail adresse is on bounce list for orga and foodsaver them self. #931 !1572 @chriswalg
 
 ## Bugfixes
 - Clarify that new forum threads won't reach members inactive for more than six months !1553 ("merging" !1385 and !1233) @zommuter